ai_assistant/requirements.txt
# AI Assistant Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# AI Provider SDKs
# =============================================================================

# Google Gemini SDK
google-generativeai>=0.8.0

# Groq SDK
groq>=0.11.0

# =============================================================================
# Web Server
# =============================================================================

# Flask web framework
flask>=3.0.0

# CORS support for Flask
flask-cors>=4.0.0

# =============================================================================
# Configuration
# =============================================================================

# Environment variable management
python-dotenv>=1.0.0
